requirements.txt
pandas
notebook
openai
python-dotenv